requirements.txt
# Bot
slack-bolt>=1.21
slack-sdk>=3.31

# LLM
anthropic>=0.40

# Database
supabase>=2.9

# WikiJS client
httpx>=0.27

# Utils
python-dotenv>=1.0

# Dev / testing
pytest>=8.0
pytest-asyncio>=0.24
ruff>=0.6
black>=24.0
